+++ C#/Libraries/SkillTools/SkillTools/SkillTools.csproj
@@ -26,6 +26,7 @@
     <DefineConstants>DEBUG;TRACE;NETFX_CORE;WINDOWS_UWP</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
+    <DocumentationFile>bin\Debug\SkillTools.XML</DocumentationFile>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
     <PlatformTarget>AnyCPU</PlatformTarget>
@@ -35,6 +36,7 @@
     <DefineConstants>TRACE;NETFX_CORE;WINDOWS_UWP</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
+    <DocumentationFile>bin\Release\SkillTools.XML</DocumentationFile>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|x86'">
     <PlatformTarget>x86</PlatformTarget>
@@ -65,6 +67,7 @@
     <DebugType>full</DebugType>
     <UseVSHostingProcess>false</UseVSHostingProcess>
     <ErrorReport>prompt</ErrorReport>
+    <DocumentationFile>bin\ARM\Debug\SkillTools.XML</DocumentationFile>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Release|ARM'">
     <PlatformTarget>ARM</PlatformTarget>
@@ -121,10 +124,10 @@
     <RestoreProjectStyle>PackageReference</RestoreProjectStyle>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="Assets\AssetWrapper.cs" />
-    <Compile Include="Assets\IAssetWrapper.cs" />
-    <Compile Include="Assets\SystemImage.cs" />
-    <Compile Include="Assets\SystemSound.cs" />
+    <Compile Include="AssetTools\AssetWrapper.cs" />
+    <Compile Include="AssetTools\IAssetWrapper.cs" />
+    <Compile Include="AssetTools\SystemImage.cs" />
+    <Compile Include="AssetTools\SystemSound.cs" />
     <Compile Include="DataStorage\EncryptedStorage.cs" />
     <Compile Include="DataStorage\ISkillStorage.cs" />
     <Compile Include="DataStorage\SecurityController.cs" />
